feat: stable version

+++ src/app/components/category-manager/category-manager.component.ts
@@ -1,5 +1,5 @@
-import { Component, inject, OnInit } from '@angular/core';
-import { ModalController } from '@ionic/angular';
+import { Component, EventEmitter, inject, OnInit, Output } from '@angular/core';
+import { AlertController, ModalController } from '@ionic/angular';
 import { LogicCoreService } from 'src/app/services/logic-core.service';
 
 @Component({
@@ -7,16 +7,23 @@ import { LogicCoreService } from 'src/app/services/logic-core.service';
   templateUrl: './category-manager.component.html',
   styleUrls: ['./category-manager.component.scss'],
 })
-export class CategoryManagerComponent  implements OnInit {
-
+export class CategoryManagerComponent implements OnInit {
+  @Output() categoryCreated = new EventEmitter<void>();
+  public logicCoreService = inject(LogicCoreService);
+  public alertController = inject(AlertController);
   public newCategoryName: string = '';
   public categories: any[] = [];
-  public logicCoreService = inject(LogicCoreService);
+  public tasks: any[] = [];
 
   constructor(private modalController: ModalController) {}
 
   ngOnInit(): void {
     this.loadCategories();
+    this.loadTasks();
+  }
+
+  loadTasks() {
+    this.tasks = this.logicCoreService.getAllTasks();
   }
 
   loadCategories() {
@@ -28,6 +35,7 @@ export class CategoryManagerComponent  implements OnInit {
       this.logicCoreService.addCategory(this.newCategoryName);
       this.newCategoryName = '';
       this.loadCategories();
+      this.categoryCreated.emit();
     }
   }
 
@@ -36,8 +44,21 @@ export class CategoryManagerComponent  implements OnInit {
     this.loadCategories();
   }
 
-  deleteCategory(categoryId: number) {
-    this.logicCoreService.deleteCategory(categoryId);
+  async deleteCategory(categoryId: number) {
+    const isCategoryUsed = this.tasks.some(task => task.categoryId === categoryId);
+
+    if (isCategoryUsed) {
+      const alert = await this.alertController.create({
+        header: 'Acción imposible',
+        message: 'No se puede eliminar esta categoría porque está asociada a una o más tareas.',
+        buttons: [],
+      });
+
+      await alert.present();
+    } else {
+      this.logicCoreService.deleteCategory(categoryId);
+    }
+
     this.loadCategories();
   }
